fix(codex): harden app-server lifecycle and permissions
Add default request timeouts, safer subprocess cleanup, and mock app-server lifecycle coverage for Codex.
Normalize permission handling and stream block indices so long sessions cannot collide or overflow.

+++ packages/service/src/domain/agents/adapter.rs
@@ -335,15 +335,15 @@ pub enum RuntimeStreamEvent {
         input_tokens: Option<u64>,
     },
     ContentBlockStart {
-        index: u32,
+        index: u64,
         block: RuntimeContentBlock,
     },
     ContentBlockDelta {
-        index: u32,
+        index: u64,
         delta: RuntimeContentDelta,
     },
     ContentBlockStop {
-        index: u32,
+        index: u64,
     },
     Other,
 }
